+++ packages/typespec-client-generator-core/src/package.ts
@@ -4,10 +4,13 @@ import {
   Diagnostic,
   getDoc,
   getNamespaceFullName,
+  getPagingOperation,
   getService,
   getSummary,
   ignoreDiagnostics,
+  isList,
   Model,
+  ModelProperty,
   Operation,
   Type,
 } from "@typespec/compiler";
@@ -18,6 +21,7 @@ import {
   getAccess,
   getClientInitialization,
   getClientNameOverride,
+  getClientNamespace,
   getOverriddenClientMethod,
   listClients,
   listOperationGroups,
@@ -42,6 +46,8 @@ import {
   SdkMethodResponse,
   SdkModelPropertyType,
   SdkModelType,
+  SdkNamespace,
+  SdkNullableType,
   SdkOperationGroup,
   SdkPackage,
   SdkPagingServiceMethod,
@@ -80,11 +86,12 @@ import {
 } from "./public-utils.js";
 import {
   addEncodeInfo,
-  getAllModelsWithDiagnostics,
+  getAllReferencedTypes,
   getClientTypeWithDiagnostics,
   getSdkCredentialParameter,
   getSdkModelPropertyType,
   getTypeSpecBuiltInType,
+  handleAllTypes,
 } from "./types.js";
 
 function getSdkServiceOperation<TServiceOperation extends SdkServiceOperation>(
@@ -112,11 +119,12 @@ function getSdkServiceOperation<TServiceOperation extends SdkServiceOperation>(
 function getSdkLroPagingServiceMethod<TServiceOperation extends SdkServiceOperation>(
   context: TCGCContext,
   operation: Operation,
+  client: SdkClientType<TServiceOperation>,
 ): [SdkLroPagingServiceMethod<TServiceOperation>, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
   return diagnostics.wrap({
-    ...diagnostics.pipe(getSdkLroServiceMethod<TServiceOperation>(context, operation)),
-    ...diagnostics.pipe(getSdkPagingServiceMethod<TServiceOperation>(context, operation)),
+    ...diagnostics.pipe(getSdkLroServiceMethod<TServiceOperation>(context, operation, client)),
+    ...diagnostics.pipe(getSdkPagingServiceMethod<TServiceOperation>(context, operation, client)),
     kind: "lropaging",
   });
 }
@@ -124,25 +132,99 @@ function getSdkLroPagingServiceMethod<TServiceOperation extends SdkServiceOperat
 function getSdkPagingServiceMethod<TServiceOperation extends SdkServiceOperation>(
   context: TCGCContext,
   operation: Operation,
+  client: SdkClientType<TServiceOperation>,
 ): [SdkPagingServiceMethod<TServiceOperation>, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
-  const pagedMetadata = getPagedResult(context.program, operation)!;
-  const basic = diagnostics.pipe(getSdkBasicServiceMethod<TServiceOperation>(context, operation));
-  if (pagedMetadata.itemsProperty) {
-    basic.response.type = diagnostics.pipe(
-      getClientTypeWithDiagnostics(context, pagedMetadata.itemsProperty.type),
+
+  const basic = diagnostics.pipe(
+    getSdkBasicServiceMethod<TServiceOperation>(context, operation, client),
+  );
+
+  // normal paging
+  if (isList(context.program, operation)) {
+    const pagingOperation = diagnostics.pipe(getPagingOperation(context.program, operation));
+
+    if (basic.response.type?.__raw?.kind !== "Model" || !pagingOperation) {
+      diagnostics.add(
+        createDiagnostic({
+          code: "unexpected-pageable-operation-return-type",
+          target: operation,
+          format: {
+            operationName: operation.name,
+          },
+        }),
+      );
+      // return as page method with no paging info
+      return diagnostics.wrap({
+        ...basic,
+        kind: "paging",
+      });
+    }
+
+    basic.response.resultPath = getPropertyPathFromModel(
+      context,
+      basic.response.type?.__raw,
+      (p) => p === pagingOperation.output.pageItems.property,
     );
+    const nextLinkPath = pagingOperation.output.nextLink
+      ? getPropertyPathFromModel(
+          context,
+          basic.response.type?.__raw,
+          (p) => p === pagingOperation.output.nextLink!.property,
+        )
+      : undefined;
+
+    context.__pagedResultSet.add(basic.response.type);
+    // tcgc will let all paging method return a list of items
+    basic.response.type = diagnostics.pipe(
+      getClientTypeWithDiagnostics(context, pagingOperation?.output.pageItems.property.type),
+    );
+
+    return diagnostics.wrap({
+      ...basic,
+      kind: "paging",
+      nextLinkPath,
+    });
   }
-  basic.response.resultPath = getPathFromSegment(
+
+  // azure core paging
+  const pagedMetadata = getPagedResult(context.program, operation)!;
+
+  if (basic.response.type?.__raw?.kind !== "Model" || !pagedMetadata.itemsProperty) {
+    diagnostics.add(
+      createDiagnostic({
+        code: "unexpected-pageable-operation-return-type",
+        target: operation,
+        format: {
+          operationName: operation.name,
+        },
+      }),
+    );
+    // return as page method with no paging info
+    return diagnostics.wrap({
+      ...basic,
+      kind: "paging",
+    });
+  }
+
+  context.__pagedResultSet.add(basic.response.type);
+
+  // tcgc will let all paging method return a list of items
+  basic.response.type = diagnostics.pipe(
+    getClientTypeWithDiagnostics(context, pagedMetadata.itemsProperty.type),
+  );
+
+  basic.response.resultPath = getPropertyPathFromSegment(
     context,
     pagedMetadata.modelType,
     pagedMetadata.itemsSegments,
   );
+
   return diagnostics.wrap({
     ...basic,
     __raw_paged_metadata: pagedMetadata,
     kind: "paging",
-    nextLinkPath: getPathFromSegment(
+    nextLinkPath: getPropertyPathFromSegment(
       context,
       pagedMetadata.modelType,
       pagedMetadata?.nextLinkSegments,
@@ -159,7 +241,45 @@ function getSdkPagingServiceMethod<TServiceOperation extends SdkServiceOperation
   });
 }
 
-function getPathFromSegment(context: TCGCContext, type: Model, segments?: string[]): string {
+export function getPropertyPathFromModel(
+  context: TCGCContext,
+  model: Model,
+  predicate: (property: ModelProperty) => boolean,
+): string | undefined {
+  const queue: { model: Model; path: ModelProperty[] }[] = [];
+
+  for (const prop of model.properties.values()) {
+    if (predicate(prop)) {
+      return getLibraryName(context, prop);
+    }
+    if (prop.type.kind === "Model") {
+      queue.push({ model: prop.type, path: [prop] });
+    }
+  }
+
+  while (queue.length > 0) {
+    const { model, path } = queue.shift()!;
+    for (const prop of model.properties.values()) {
+      if (predicate(prop)) {
+        return path
+          .concat(prop)
+          .map((s) => getLibraryName(context, s))
+          .join(".");
+      }
+      if (prop.type.kind === "Model") {
+        queue.push({ model: prop.type, path: path.concat(prop) });
+      }
+    }
+  }
+
+  return undefined;
+}
+
+function getPropertyPathFromSegment(
+  context: TCGCContext,
+  type: Model,
+  segments?: string[],
+): string {
   if (!segments || segments.length === 0) {
     return "";
   }
@@ -179,11 +299,12 @@ function getPathFromSegment(context: TCGCContext, type: Model, segments?: string
 function getSdkLroServiceMethod<TServiceOperation extends SdkServiceOperation>(
   context: TCGCContext,
   operation: Operation,
+  client: SdkClientType<TServiceOperation>,
 ): [SdkLroServiceMethod<TServiceOperation>, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
   const metadata = getServiceMethodLroMetadata(context, operation)!;
   const basicServiceMethod = diagnostics.pipe(
-    getSdkBasicServiceMethod<TServiceOperation>(context, operation),
+    getSdkBasicServiceMethod<TServiceOperation>(context, operation, client),
   );
 
   basicServiceMethod.response.type = metadata.finalResponse?.result;
@@ -245,6 +366,7 @@ function getSdkMethodResponse(
   context: TCGCContext,
   operation: Operation,
   sdkOperation: SdkServiceOperation,
+  client: SdkClientType<SdkServiceOperation>,
 ): SdkMethodResponse {
   const responses = sdkOperation.responses;
   // TODO: put head as bool here
@@ -256,9 +378,12 @@ function getSdkMethodResponse(
     type = {
       __raw: operation,
       kind: "union",
+      access: "public",
+      usage: UsageFlags.Output,
       variantTypes: allResponseBodies,
       name: createGeneratedName(context, operation, "UnionResponse"),
       isGeneratedName: true,
+      clientNamespace: client.clientNamespace,
       crossLanguageDefinitionId: getCrossLanguageDefinitionId(context, operation),
       decorators: [],
     };
@@ -270,6 +395,9 @@ function getSdkMethodResponse(
       kind: "nullable",
       type: type,
       decorators: [],
+      access: "public",
+      usage: UsageFlags.Output,
+      clientNamespace: client.clientNamespace,
     };
   }
   return {
@@ -281,6 +409,7 @@ function getSdkMethodResponse(
 function getSdkBasicServiceMethod<TServiceOperation extends SdkServiceOperation>(
   context: TCGCContext,
   operation: Operation,
+  client: SdkClientType<TServiceOperation>,
 ): [SdkServiceMethod<TServiceOperation>, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
   const methodParameters: SdkMethodParameter[] = [];
@@ -339,7 +468,7 @@ function getSdkBasicServiceMethod<TServiceOperation extends SdkServiceOperation>
       addEncodeInfo(context, methodBodyParam.__raw!, methodBodyParam.type, defaultContentType),
     );
   }
-  const response = getSdkMethodResponse(context, operation, serviceOperation);
+  const response = getSdkMethodResponse(context, operation, serviceOperation, client);
   const name = getLibraryName(context, operation);
   return diagnostics.wrap({
     __raw: operation,
@@ -366,23 +495,25 @@ function getSdkBasicServiceMethod<TServiceOperation extends SdkServiceOperation>
     decorators: diagnostics.pipe(getTypeDecorators(context, operation)),
     generateConvenient: shouldGenerateConvenient(context, operation),
     generateProtocol: shouldGenerateProtocol(context, operation),
+    isOverride: override !== undefined,
   });
 }
 
 function getSdkServiceMethod<TServiceOperation extends SdkServiceOperation>(
   context: TCGCContext,
   operation: Operation,
+  client: SdkClientType<TServiceOperation>,
 ): [SdkServiceMethod<TServiceOperation>, readonly Diagnostic[]] {
   const lro = getLroMetadata(context.program, operation);
-  const paging = getPagedResult(context.program, operation);
+  const paging = getPagedResult(context.program, operation) || isList(context.program, operation);
   if (lro && paging) {
-    return getSdkLroPagingServiceMethod<TServiceOperation>(context, operation);
+    return getSdkLroPagingServiceMethod<TServiceOperation>(context, operation, client);
   } else if (paging) {
-    return getSdkPagingServiceMethod<TServiceOperation>(context, operation);
+    return getSdkPagingServiceMethod<TServiceOperation>(context, operation, client);
   } else if (lro) {
-    return getSdkLroServiceMethod<TServiceOperation>(context, operation);
+    return getSdkLroServiceMethod<TServiceOperation>(context, operation, client);
   }
-  return getSdkBasicServiceMethod<TServiceOperation>(context, operation);
+  return getSdkBasicServiceMethod<TServiceOperation>(context, operation, client);
 }
 
 function getClientDefaultApiVersion(
@@ -430,6 +561,7 @@ function getSdkInitializationType(
       access,
       usage: UsageFlags.Input,
       crossLanguageDefinitionId: `${getNamespaceFullName(client.service.namespace!)}.${name}`,
+      clientNamespace: getClientNamespace(context, client.type),
       apiVersions: context.__tspTypeToApiVersions.get(client.type)!,
       decorators: [],
     };
@@ -481,7 +613,9 @@ function getSdkMethods<TServiceOperation extends SdkServiceOperation>(
   const diagnostics = createDiagnosticCollector();
   const retval: SdkMethod<TServiceOperation>[] = [];
   for (const operation of listOperationsInOperationGroup(context, client)) {
-    retval.push(diagnostics.pipe(getSdkServiceMethod<TServiceOperation>(context, operation)));
+    retval.push(
+      diagnostics.pipe(getSdkServiceMethod<TServiceOperation>(context, operation, sdkClientType)),
+    );
   }
   for (const operationGroup of listOperationGroups(context, client)) {
     // We create a client accessor for each operation group
@@ -620,10 +754,13 @@ function getSdkEndpointParameter<TServiceOperation extends SdkServiceOperation =
   if (types.length > 1) {
     type = {
       kind: "union",
+      access: "public",
+      usage: UsageFlags.None,
       variantTypes: types,
       name: createGeneratedName(context, rawClient.service, "Endpoint"),
       isGeneratedName: true,
       crossLanguageDefinitionId: getCrossLanguageDefinitionId(context, rawClient.service),
+      clientNamespace: getClientNamespace(context, rawClient.service),
       decorators: [],
     } as SdkUnionType<SdkEndpointType>;
   } else {
@@ -667,6 +804,7 @@ function createSdkClientType<TServiceOperation extends SdkServiceOperation>(
     methods: [],
     apiVersions: context.__tspTypeToApiVersions.get(client.type)!,
     nameSpace: getClientNamespaceStringHelper(context, client.service)!,
+    clientNamespace: getClientNamespace(context, client.type),
     initialization: diagnostics.pipe(getSdkInitializationType(context, client)),
     decorators: diagnostics.pipe(getTypeDecorators(context, client.type)),
     parent,
@@ -760,14 +898,75 @@ export function getSdkPackage<TServiceOperation extends SdkServiceOperation>(
 ): [SdkPackage<TServiceOperation>, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
   populateApiVersionInformation(context);
-  const modelsAndEnums = diagnostics.pipe(getAllModelsWithDiagnostics(context));
+  diagnostics.pipe(handleAllTypes(context));
   const crossLanguagePackageId = diagnostics.pipe(getCrossLanguagePackageId(context));
-  return diagnostics.wrap({
+  const allReferencedTypes = getAllReferencedTypes(context);
+  const sdkPackage: SdkPackage<TServiceOperation> = {
     name: getClientNamespaceString(context)!,
     rootNamespace: getClientNamespaceString(context)!,
     clients: listClients(context).map((c) => diagnostics.pipe(createSdkClientType(context, c))),
-    models: modelsAndEnums.filter((x): x is SdkModelType => x.kind === "model"),
-    enums: modelsAndEnums.filter((x): x is SdkEnumType => x.kind === "enum"),
+    models: allReferencedTypes.filter((x): x is SdkModelType => x.kind === "model"),
+    enums: allReferencedTypes.filter((x): x is SdkEnumType => x.kind === "enum"),
+    unions: allReferencedTypes.filter(
+      (x): x is SdkUnionType | SdkNullableType => x.kind === "union" || x.kind === "nullable",
+    ),
     crossLanguagePackageId,
-  });
+    namespaces: [],
+  };
+  organizeNamespaces(sdkPackage);
+  return diagnostics.wrap(sdkPackage);
+}
+
+function organizeNamespaces<TServiceOperation extends SdkServiceOperation>(
+  sdkPackage: SdkPackage<TServiceOperation>,
+) {
+  const clients = [...sdkPackage.clients];
+  while (clients.length > 0) {
+    const client = clients.shift()!;
+    getSdkNamespace(sdkPackage, client.clientNamespace).clients.push(client);
+    client.methods
+      .filter((m) => m.kind === "clientaccessor")
+      .map((m) => m.response)
+      .map((c) => clients.push(c));
+  }
+  for (const model of sdkPackage.models) {
+    getSdkNamespace(sdkPackage, model.clientNamespace).models.push(model);
+  }
+  for (const enumType of sdkPackage.enums) {
+    getSdkNamespace(sdkPackage, enumType.clientNamespace).enums.push(enumType);
+  }
+  for (const unionType of sdkPackage.unions) {
+    getSdkNamespace(sdkPackage, unionType.clientNamespace).unions.push(unionType);
+  }
+}
+
+function getSdkNamespace<TServiceOperation extends SdkServiceOperation>(
+  sdkPackage: SdkPackage<TServiceOperation>,
+  namespace: string,
+) {
+  const segments = namespace.split(".");
+  let current: SdkPackage<TServiceOperation> | SdkNamespace<TServiceOperation> = sdkPackage;
+  let fullName = "";
+  for (const segment of segments) {
+    fullName = fullName === "" ? segment : `${fullName}.${segment}`;
+    const ns: SdkNamespace<TServiceOperation> | undefined = current.namespaces.find(
+      (ns) => ns.name === segment,
+    );
+    if (ns === undefined) {
+      const newNs = {
+        name: segment,
+        fullName,
+        clients: [],
+        models: [],
+        enums: [],
+        unions: [],
+        namespaces: [],
+      };
+      current.namespaces.push(newNs);
+      current = newNs;
+    } else {
+      current = ns;
+    }
+  }
+  return current;
 }