+++ packages/azure-http-specs/CHANGELOG.md
@@ -1,1 +1,9 @@
 # @azure-tools/azure-http-specs
+
+## 0.1.0-alpha.2
+
+Backport changes from `cadl-ranch`.
+
+## 0.1.0-alpha.1
+
+No changes, version bump only.